Changed __proxy__ to prepare eagerly.
Previously, the proxy class was prepared lazily (excuse the pun):

    lazy_identity = lazy(identity, int)
    lazy_identity(10)  # <- prepared here
    lazy_identity(10)

This has a slight advantage that if the lazy doesn't end up getting
used, the preparation work is skipped, however that's not very likely.

Besides this laziness, it is also inconsistent in that the methods which
are wrapped directly (__str__ etc.) are prepared already when __proxy__
is defined, and there is a weird half-initialized state.

Change it so that everything is prepared already on the first line of
the example above.